% Generated by roxygen2 (4.0.1): do not edit by hand
\name{example_function}
\alias{example_function}
\title{An example function that sums two numbers}
\usage{
example_function(a, b)
}
\arguments{
\item{a}{a number}

\item{b}{a number}
}
\value{
a numeric vector giving the component-wise sum.
}
\description{
An example function that sums two numbers
}
\examples{
example_function(1,1)
example_function(pi,exp(1))
example_function(1:2,3:4)
}
\seealso{
'+'
}

