% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{fluTrends}
\alias{fluTrends}
\title{Google fluTrends}
\format{A data frame with 583 rows and 160 columns}
\usage{
data(fluTrends)
}
\description{
A data set containing google flu trends data (<http://www.google.org/flutrends/us/#US>)
for the United States.
Downloaded from <http://www.google.org/flutrends/us/data.txt> on 26 Nov
2014.
}
\details{
For each column, an index of the current influenza activity level
for that region. The index is based on google search engine queries.
}
\author{
Jarad Niemi
\itemize{
  \item Date Date. The week of the observations
}
}
\references{
Ginsberg, Jeremy, Matthew H. Mohebbi, Rajan S. Patel, Lynnette Brammer, Mark S. Smolinski, and Larry Brilliant. "Detecting influenza epidemics using search engine query data." Nature 457, no. 7232 (2008): 1012-1014.
}
\keyword{datasets}

