% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GI}
\alias{GI}
\title{Fictional syndromic surveillance data set}
\format{A data frame with 21244 rows and 8 columns}
\description{
A data set containing fictitious patient data for gastrointestinal (GI)
visits to an emergency department (ER).
}
\details{
\itemize{
  \item id. patient id
  \item date. date of visit to ER
  \item facility. facility id
  \item icd9. international statistical classificiation of diseases, ninth revision, code
  \item age. patient age
  \item zipcode. patient zip code
  \item chief_complain. free text chief complaint
  \item syndrome. syndromic surveillance code
}
}

