% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd9.R
\docType{data}
\name{icd9}
\alias{icd9}
\title{ICD9 codes}
\format{A data frame with 19 rows and 2 columns}
\usage{
data(icd9)
}
\description{
A data set that contains the International Statistical Classification of Diseases 
and Related Health Problems (ICD-9) codes and their classification.
}
\author{
Jarad Niemi 
\itemize{
  \item code_cutpoint The lower end of the range for the ICD9 code representing that classification.
  \item classification The ICD9 classification
}
}
\keyword{datasets}

