% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_additional_packages.R
\name{install_additonal_packages}
\alias{install_additonal_packages}
\title{A function that installs all supplementary packages for the workshop.}
\usage{
install_additonal_packages()
}
\description{
A function that installs all supplementary packages for the workshop.
}
\details{
This function installs all additional packages that are used
throughout the workshop, but are not absolutely necessary for the workshop.
}
\examples{
\dontrun{
install_additional_packages()
}
}
\seealso{
\code{\link{MWBDSSworkshop}}
}
\author{
Jarad Niemi <\url{http://jarad.me}>
}
