\name{ISDSWorkshop-package}
\alias{ISDSWorkshop-package}
\alias{ISDSWorkshop}
\docType{package}
\title{
Vignettes and data sets for the ISDS R workshop. 
}
\description{
This package contains vignettes and data sets for the International Society for Disease surveillance R workshop. 
}
\details{
\tabular{ll}{
Package: \tab ISDSWorkshop\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-30\cr
License: \tab GPL\cr
}
vignette(index)
}
\author{
Jarad Niemi

Maintainer: Jarad Niemi <niemi@iastate.edu>
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{
}
