% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MWBDSSworkshop-package}
\alias{MWBDSSworkshop-package}
\alias{MWBDSSworkshop}
\title{An R package for the Midwest Big Data Summer School R Workshop.}
\description{
This is intended to be a stand-alone package for all the materials in this training.
The 'slides' are available as HTML vignettes which contains the code, its output, and 
supporting text.
}
\seealso{
To launch the workshop, use \code{\link{workshop}}.
}
\author{
Jarad Niemi <\url{http://jarad.me}>
}
