% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{ISDSWorkshop-package}
\alias{ISDSWorkshop}
\alias{ISDSWorkshop-package}
\title{An R package for the ISDS Introduction to R for Biosurveillance pre-conference training.}
\description{
This is intended to be a stand-alone package for all the materials in this training.
The 'slides' are available as HTML vignettes which contains the code, its output, and
supporting text.
}
\author{
Jarad Niemi <\url{http://jarad.me}>
}
\seealso{
To launch the workshop, use \code{\link{workshop}}.
}

