% Generated by roxygen2 (4.0.2): do not edit by hand
\name{workshop}
\alias{workshop}
\title{A function that prepares the workshop.}
\usage{
workshop(write_data = TRUE, clean = TRUE, launch_index = TRUE)
}
\arguments{
\item{write_data}{logical. If TRUE, writes the GI data set to a csv file.}

\item{clean}{logical. If TRUE, removes all objects created by the function from the R workspace.}

\item{launch_index}{logical. If TRUE, launches the workshop html index.}
}
\description{
A function that prepares the workshop.
}
\examples{
workshop()
 \dontrun{
 workshop(write_data = FALSE)
 workshop(clean = FALSE)
 }
}
\author{
Jarad Niemi <\url{http://jarad.me}>
}
\seealso{
The function uses the \code{\link{data}} function to load the \code{\link{GI}} data set and
   then uses the \code{\link{write.csv}} function to write it to a file.
   It then launches the \code{\link{ISDSWorkshop}} html index
   using the \code{\link{vignette}} function.
}

